+++ pkg/analysis_server/lib/src/services/refactoring/move_file.dart
@@ -89,8 +89,8 @@ class MoveFileRefactoringImpl extends RefactoringImpl
       if (libraryResult is! ResolvedLibraryResult) {
         return changeBuilder.sourceChange;
       }
-      var definingUnitResult = libraryResult.units!.first;
-      for (var result in libraryResult.units!) {
+      var definingUnitResult = libraryResult.units.first;
+      for (var result in libraryResult.units) {
         if (result.isPart) {
           var partOfs = result.unit!.directives
               .whereType<PartOfDirective>()