+++ pkg/nnbd_migration/lib/src/front_end/info_builder.dart
@@ -98,7 +98,7 @@ class InfoBuilder {
       var session = driverProvider!.getAnalysisSession(filePath);
       var result = await session.getResolvedLibrary2(filePath!);
       if (result is ResolvedLibraryResult) {
-        for (var unitResult in result.units!) {
+        for (var unitResult in result.units) {
           var sourceInfo =
               sourceInfoMap[unitResult.unit!.declaredElement!.source];
           // Note: there might have been no information for this unit in